Move staleness tracking out of the database layer
TxCursor now owns a StaleTracker (serviceshared/matching/tracker.py) and
reports what it is about to run, what it ran, and what it fetched; every
piece of model knowledge -- classification, capture reads, attribution,
harvest, the pre-commit flush -- lives with the models. Adding a model
means adding a module under serviceshared/matching and appending it to
MODELS; the database layer never changes.

The move also closes a latent hole: a captured-table write through
executemany used to mark people stale without any captured old values,
which a model would treat as nothing-to-do. note_after now requires the
capture to have actually happened, so that shape raises instead of
silently under-refreshing.
